refactor: consolidate config management into local commands
Move config init functionality from separate config module to local
commands to simplify CLI structure. Remove standalone config command
and integrate template-based initialization directly into local
workflow. Update admin clean command to reference new location.
